classify logical sort into a separate file for later pkg move
