+++ pkg/planner/core/rule_topn_push_down.go
@@ -65,23 +65,6 @@ func (p *LogicalCTE) PushDownTopN(topNLogicalPlan base.LogicalPlan, opt *optimiz
 	return p
 }
 
-// PushDownTopN implements LogicalPlan interface.
-func (ls *LogicalSort) PushDownTopN(topNLogicalPlan base.LogicalPlan, opt *optimizetrace.LogicalOptimizeOp) base.LogicalPlan {
-	var topN *LogicalTopN
-	if topNLogicalPlan != nil {
-		topN = topNLogicalPlan.(*LogicalTopN)
-	}
-	if topN == nil {
-		return ls.BaseLogicalPlan.PushDownTopN(nil, opt)
-	} else if topN.isLimit() {
-		topN.ByItems = ls.ByItems
-		appendSortPassByItemsTraceStep(ls, topN, opt)
-		return ls.Children()[0].PushDownTopN(topN, opt)
-	}
-	// If a TopN is pushed down, this sort is useless.
-	return ls.Children()[0].PushDownTopN(topN, opt)
-}
-
 // PushDownTopN implements the LogicalPlan interface.
 func (p *LogicalUnionAll) PushDownTopN(topNLogicalPlan base.LogicalPlan, opt *optimizetrace.LogicalOptimizeOp) base.LogicalPlan {
 	var topN *LogicalTopN
